Move features element to gradle module.
