+++ feature_bone/src/main/java/com/hfad/palamarchuksuperapp/feature/bone/ui/OrderCard.kt
@@ -1,4 +1,4 @@
-package com.hfad.palamarchuksuperapp.ui.compose.boneScreen
+package com.hfad.palamarchuksuperapp.feature.bone.ui
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.animateDpAsState
@@ -44,15 +44,15 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.hfad.palamarchuksuperapp.core.ui.theme.AppTheme
-import com.hfad.palamarchuksuperapp.R
 import com.hfad.palamarchuksuperapp.core.ui.composables.basic.AppText
-import com.hfad.palamarchuksuperapp.ui.compose.utils.StepProgressionBar
 import com.hfad.palamarchuksuperapp.core.ui.composables.basic.appTextConfig
-import com.hfad.palamarchuksuperapp.ui.viewModels.Order
-import com.hfad.palamarchuksuperapp.ui.viewModels.ServiceType
-import com.hfad.palamarchuksuperapp.ui.viewModels.StepperStatus
-import com.hfad.palamarchuksuperapp.ui.viewModels.orderServiceList
+import com.hfad.palamarchuksuperapp.core.ui.theme.AppTheme
+import com.hfad.palamarchuksuperapp.feature.bone.R
+import com.hfad.palamarchuksuperapp.feature.bone.ui.composables.StepProgressionBar
+import com.hfad.palamarchuksuperapp.feature.bone.ui.viewModels.Order
+import com.hfad.palamarchuksuperapp.feature.bone.ui.viewModels.ServiceType
+import com.hfad.palamarchuksuperapp.feature.bone.ui.viewModels.StepperStatus
+import com.hfad.palamarchuksuperapp.feature.bone.ui.viewModels.orderServiceList
 
 @Composable
 fun OrderCard(
